requirements.txt
beautifulsoup4
ipykernel
MLB-StatsAPI
pandas
requests